feat: flexible database labels
