--- frontend/src/components/DatabaseLabels/index.ts
@@ -1,3 +1,0 @@
-import DatabaseLabelProps from "./DatabaseLabelProps.vue";
-
-export { DatabaseLabelProps };
