+++ frontend/src/components/DeploymentConfigTool/DeploymentMatrix.vue
@@ -41,7 +41,7 @@
 
 <script lang="ts" setup>
 import { computed, reactive } from "vue";
-import { ComposedDatabase, LabelKeyType } from "@/types";
+import { ComposedDatabase } from "@/types";
 import { Environment } from "@/types/proto/v1/environment_service";
 import { DeploymentConfig } from "@/types/proto/v1/project_service";
 import { filterDatabaseV1ByKeyword } from "@/utils";
@@ -60,7 +60,7 @@ const props = withDefaults(
 );
 
 const state = reactive({
-  label: "environment" as LabelKeyType,
+  label: "environment",
   keyword: "",
 });
